+++ src/onegov.election_day/onegov/election_day/views/manage/elections.py
@@ -4,6 +4,7 @@ import morepath
 
 from onegov.ballot import Election, ElectionCollection
 from onegov.core.security import Private
+from onegov.core.utils import groupbylist
 from onegov.election_day import _
 from onegov.election_day import ElectionDayApp
 from onegov.election_day.collections import ArchivedResultCollection
@@ -21,7 +22,7 @@ def view_elections(self, request):
     return {
         'layout': ManageElectionsLayout(self, request),
         'title': _("Manage"),
-        'elections': self.batch,
+        'groups': groupbylist(self.batch, key=lambda item: item.date),
         'new_election': request.link(self, 'new-election')
     }